Migrated the example package to null safety
