+++ example/pubspec.yaml
@@ -3,9 +3,10 @@ description: Showcases the use of Material Segmented Control widget
 version: 1.0.0
 author: The.BENN<beagle-barks-contact@yahoo.com>
 homepage: https://github.com/beagle-barks/material_segmented_control
+publish_to: 'none'
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0-29.10.beta <3.0.0'
 
 dependencies:
   flutter:
